requirements.txt
PyQt5
pdfminer
filetype
requests
PyExecJS
sip